Properly resize captured image before displaying it.
- scales up or down the display image depending on the video and server sizes
- adds server_flip argument to get_html that allows sending images in their natural left-right orientation

Flipping image makes perfect sense for displaying on the screen (like a mirror), but depending on your use case you might prefer an image as it would be seen by a person standing where the webcam is. Default is to send the same image that the user sees on the screen (flipped).